Add evm fee calculator tests

+++ blockchain/src/main/java/com/gemwallet/android/blockchain/clients/ethereum/EvmSignClient.kt
@@ -77,7 +77,7 @@ class EvmSignClient(
             amount = amount,
             tokenAmount = params.finalAmount,
             fee = fee,
-            chainId = meta.chainId,
+            chainId = meta.chainId.toBigInteger(),
             nonce = meta.nonce,
             destinationAddress = params.input.destination()?.address ?: "",
             memo = when (input) {
